LRAC-12616 Create '__tests__' folder to Properties and move files to it
